+++ MonoTanksBotLogic/Networking/Packets/gameState/Zone.cs
@@ -13,4 +13,4 @@ namespace MonoTanksClientLogic.Models;
 /// <param name="Height">Represents zone height.</param>
 /// <param name="Status">Represents zone capture status.</param>
 [JsonConverter(typeof(ZoneJsonConverter))]
-public record class Zone(int Index, long X, long Y, long Width, long Height, ZoneStatus Status);
+public record class Zone(int Index, int X, int Y, int Width, int Height, ZoneStatus Status);
